+++ lib/browser/mixins/browser_click.ts
@@ -1,7 +1,8 @@
 import BrowserActions from './browser_actions';
-import { WendigoError, QueryError } from '../errors';
-import DomElement from '../models/dom_element';
-import { WendigoSelector } from '../types';
+
+import { WendigoError, QueryError } from '../../errors';
+import DomElement from '../../models/dom_element';
+import { WendigoSelector } from '../../types';
 
 export default abstract class BrowserClick extends BrowserActions {
     public async click(selector: WendigoSelector | number, index?: number): Promise<number> {