+++ lib/browser/mixins/browser_info.ts
@@ -1,6 +1,7 @@
 import BrowserClick from './browser_click';
-import { QueryError } from '../errors';
-import { WendigoSelector } from '../types';
+
+import { QueryError } from '../../errors';
+import { WendigoSelector } from '../../types';
 
 export default abstract class BrowserInfo extends BrowserClick {
     public text(selector: WendigoSelector): Promise<Array<string>> {