+++ lib/browser/mixins/browser_wait.ts
@@ -1,10 +1,11 @@
 import { EvaluateFn } from 'puppeteer';
 
-import * as utils from '../utils/utils';
-import DomElement from '../models/dom_element';
-import { TimeoutError, WendigoError } from '../errors';
 import BrowserNavigation from './browser_navigation';
-import { CssSelector, WendigoSelector } from '../types';
+
+import * as utils from '../../utils/utils';
+import DomElement from '../../models/dom_element';
+import { TimeoutError, WendigoError } from '../../errors';
+import { CssSelector, WendigoSelector } from '../../types';
 
 export default abstract class BrowserWait extends BrowserNavigation {
     public wait(ms: number = 250): Promise<void> {